feat: refactor tests
