+++ tests/test-utils.test.ts
@@ -1,4 +1,5 @@
 import { EditorState, basicSetup, EditorView } from "@codemirror/basic-setup";
+import { Extension } from "@codemirror/state";
 import { vim } from "@replit/codemirror-vim";
 import { blockLevelExtension } from "../src/extension";
 
@@ -6,6 +7,19 @@ export function createTestEditorWithDoc(content: string) {
   let initialState = EditorState.create({
     doc: content,
     extensions: [vim(), blockLevelExtension(), basicSetup],
+  });
+  return new EditorView({
+    state: initialState,
+  });
+}
+
+export function createTestEditorWithExtensionsAndDoc(
+  extensions: Extension[],
+  content: string
+) {
+  let initialState = EditorState.create({
+    doc: content,
+    extensions,
   });
   return new EditorView({
     state: initialState,
